requirements.txt
python-telegram-bot==20.6
pymongo==4.5.0
python-dotenv==1.0.0
motor==3.3.1
asyncio==3.4.3
tzlocal==5.0.1 